+++ tests/testthat/test-plot-trajectory.R
@@ -1,5 +1,3 @@
-context("plot.trajectory")
-
 test_graph <- function(x, name, from, to) {
   expect_true(inherits(plot(x), "htmlwidget"))
   graph <- trajectory_graph(x, scales::brewer_pal("qual"))
@@ -39,6 +37,21 @@ test_that("a rollback with variable amount is correctly converted to graph", {
   test_graph(trajectory() %>% timeout(1) %>% rollback(1), c("Timeout", "Rollback"), c(1, 2), c(2, 1))
   test_graph(trajectory() %>% timeout(1) %>% rollback(2), c("Timeout", "Rollback"), c(1, 2), c(2, 1))
   test_graph(trajectory() %>% timeout(1) %>% rollback(Inf), c("Timeout", "Rollback"), c(1, 2), c(2, 1))
+})
+
+test_that("a rollback with tag is correctly converted to graph", {
+  skip_if(packageVersion("simmer") <= "4.4.5")
+
+  x <- trajectory() %>%
+    timeout(1, tag="foo") %>%
+    timeout(2, tag="foo") %>%
+    timeout(3) %>%
+    rollback("foo", tag="bar")
+
+  test_graph(x,
+             c("Timeout [foo]", "Timeout [foo]", "Timeout", "Rollback [bar]"),
+             c(1, 2, 3, 4),
+             c(2, 3, 4, 2))
 })
 
 test_that("a complex trajectory is correctly converted to graph", {
@@ -104,4 +117,27 @@ test_that("edges are removed for ignored subtrajectories in Clone", {
                "Clone", "Timeout", "SetAttribute", "Clone", "Timeout", "SetAttribute", "Synchronize"),
              c(1, 1, 1, 2, 3, 4, 4, 5, 6, 7, 8, 9, 11, 12),
              c(2, 3, 4, 4, 4, 5, 6, 7, 7, 8, 10, 10, 13, 13))
+})
+
+test_that("edges are not postprocessed when n is specified as a function", {
+  x <- trajectory() %>%
+    clone(function() 3,
+          trajectory() %>% timeout(1),
+          trajectory() %>% set_attribute("asdf", 1)) %>%
+    clone(function() 2,
+          trajectory() %>% timeout(1),
+          trajectory() %>% set_attribute("asdf", 1)) %>%
+    clone(function() 1,
+          trajectory() %>% timeout(1),
+          trajectory() %>% set_attribute("asdf", 1)) %>%
+    clone(function() 0,
+          trajectory() %>% timeout(1),
+          trajectory() %>% set_attribute("asdf", 1)) %>%
+    synchronize()
+
+  test_graph(x,
+             c("Clone", "Timeout", "SetAttribute", "Clone", "Timeout", "SetAttribute",
+               "Clone", "Timeout", "SetAttribute", "Clone", "Timeout", "SetAttribute", "Synchronize"),
+             c(1, 1, 1, 2, 3, 4, 4, 4, 5, 6, 7, 7, 7, 8, 9, 10, 10, 10, 11, 12),
+             c(2, 3, 4, 4, 4, 5, 6, 7, 7, 7, 8, 9, 10, 10, 10, 11, 12, 13, 13, 13))
 })